+++ src/Models/ObjectMap/Entities/EntityGubbins.php
@@ -119,13 +119,13 @@ class EntityGubbins
     public function setFields(array $fields)
     {
         if (0 == count($fields)) {
-            $msg = 'Fields array must not be empty';
+            $msg = 'Fields array must not be empty for '.$this->getClassName();
             throw new \Exception($msg);
         }
         $keys = [];
         foreach ($fields as $propName => $field) {
             if (!$field instanceof EntityField) {
-                $msg = 'Fields array must only have EntityField objects';
+                $msg = 'Fields array must only have EntityField objects for '.$this->getClassName();
                 throw new \Exception($msg);
             }
             if ($field->getIsKeyField()) {
@@ -133,7 +133,7 @@ class EntityGubbins
             }
         }
         if (0 == count($keys)) {
-            $msg = 'No key field supplied in fields array';
+            $msg = 'No key field supplied in fields array for '.$this->getClassName();
             throw new \Exception($msg);
         }
         $this->fields = $fields;